Rename pyrocky package path
Package path renamed to follow PyAnsys standards.

ROCKY-20686
