+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "ansys-rocky-prepost"
+name = "ansys-rocky-core"
 version = "0.4.0"
 authors = [
   { name="ANSYS, Inc.", email="pyansys.support@ansys.com" },
