+++ src/ansys/rocky/core/exceptions.py
